refactor(frontend): maintain a dedicated logout page to separate session termination from navigation handling, preventing related issues

+++ modules/core/handler_modules.php
@@ -778,9 +778,18 @@ class Hm_Handler_logout extends Hm_Handler_Module {
      * Clean up everything on logout
      */
     public function process() {
+        if ($this->request->get['prompt'] ?? false) {
+            $backQuery = isset($this->request->get['back_query']) ? unserialize(base64_decode($this->request->get['back_query'])): [];
+
+            $this->out('cancel_logout_url', '?' . http_build_query($backQuery));
+            
+            return;
+        }
+
         if (array_key_exists('logout', $this->request->post) && !$this->session->loaded) {
             $this->session->destroy($this->request);
             Hm_Msgs::add('Session destroyed on logout', 'info');
+            $this->out('redirect_url', '?home');
         }
         elseif (array_key_exists('save_and_logout', $this->request->post)) {
             list($success, $form) = $this->process_form(array('password'));
@@ -803,6 +812,7 @@ class Hm_Handler_logout extends Hm_Handler_Module {
                         $this->user_config->save($user, $pass);
                         $this->session->destroy($this->request);
                         Hm_Msgs::add('Saved user data on logout, Session destroyed on logout', 'info');
+                        $this->out('redirect_url', '?home');
                     } catch (Exception $e) {
                         Hm_Msgs::add('Could not save settings: ' . $e->getMessage(), 'warning');
                     }
